+++ env.go
@@ -7,13 +7,6 @@ func Env(k, v string) Stubber {
 	return newStubs().Env(k, v)
 }
 
-func (s *stubs) checkEnvKey(k string) {
-	if _, ok := s.envs[k]; !ok {
-		v, ok := os.LookupEnv(k)
-		s.envs[k] = env{v, ok}
-	}
-}
-
 // Env the specified environent variable to the specified value.
 func (s *stubs) Env(k, v string) Stubber {
 	s.checkEnvKey(k)
@@ -21,10 +14,9 @@ func (s *stubs) Env(k, v string) Stubber {
 	return s
 }
 
-// UnsetEnv unset the specified environent variable.
-func (s *stubs) UnsetEnv(k string) *stubs {
-	s.checkEnvKey(k)
-
-	os.Unsetenv(k)
-	return s
+func (s *stubs) checkEnvKey(k string) {
+	if _, ok := s.envs[k]; !ok {
+		v, existing := os.LookupEnv(k)
+		s.envs[k] = env{v, existing}
+	}
 }