+++ env_test.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"testing"
 
+	. "github.com/smartystreets/goconvey/convey"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -31,4 +32,25 @@ func TestStubEnv(t *testing.T) {
 
 	assert.Equal(t, "V1", os.Getenv("GOSTUB_T1"), "Wrong reset value for T1")
 	assert.Equal(t, "V2", os.Getenv("GOSTUB_T2"), "Wrong reset value for T2")
+}
+
+func TestEnv(t *testing.T) {
+	k, v := "TestEnv", "true"
+	Convey("如果想要打桩了环境变量 \"TestEnv\" 为 \"true\"", t, func() {
+		Convey("在打桩前，该环境变量应该不存在", func() {
+			_, ok := os.LookupEnv(k)
+			So(ok, ShouldBeFalse)
+		})
+		stubs := Env(k, v)
+		Convey("打桩后，可以查询到该环境变量", func() {
+			actual, ok := os.LookupEnv(k)
+			So(ok, ShouldBeTrue)
+			So(actual, ShouldEqual, v)
+		})
+		stubs.Restore()
+		Convey("恢复后，该环境变量被删除", func() {
+			_, ok := os.LookupEnv(k)
+			So(ok, ShouldBeFalse)
+		})
+	})
 }